Migrate some state tests into pytest
